feat(scan): sub-pixel tool trace from the soft confidence mask (#2240)
The tool outline was traced by boundary-walking a *binary* mask: the
MediaPipe segmenter's soft output was thresholded to 0/1, then every
contour vertex snapped to an integer pixel. That bakes in a ±0.5px
staircase on every edge (~±0.05-0.1mm at the card's ground scale) and
throws away the sub-pixel boundary the model already encodes.

Request MediaPipe's confidence mask (a [0,1] field, on by default; category
mask kept as a fallback) and trace its 0.5 iso-contour with marching squares
that interpolates each vertex to where the probability actually crosses 0.5.
Outlines come out smooth instead of stair-stepped.

- new softContour.ts: marching-squares tracer (largest closed loop, wound
  clockwise to match traceContour) + binarize() for consumers that need 0/1.
- traceScene: shared finishTrace tail; buildToolTraceSoft for the ML path;
  traceSceneSegmented takes the soft mask, binarizing only for card exclusion.
- segmentAt returns a SoftMask; ScanPage threads the type through.

Downstream simplify/smooth/rectify and the card homography are unchanged.
This is step 1 of the trace-accuracy work; guided-filter edge-snapping and
shape regularization (symmetry, arc fitting) follow.

+++ src/features/scan-capture/components/ScanPage/ScanPage.test.tsx
@@ -31,7 +31,7 @@ vi.mock('@/i18n', () => ({
 
 const TOKEN = '3f2504e0-4f89-41d3-9a0c-0305e82c3301';
 const IMAGE = { width: 24, height: 20, data: new Uint8ClampedArray(24 * 20 * 4) };
-const MASK = { width: 24, height: 20, data: new Uint8Array(24 * 20) };
+const MASK = { width: 24, height: 20, data: new Float32Array(24 * 20) };
 const PTS = [
   { x: 0, y: 0 },
   { x: 25, y: 0 },
